Work on tenant provisioning
This creates a state machine that does tenant provisioning from the management account. Each tenant shard has a provisioning role, that can be assumed by the management account, and a table for recording tenant info (to be used by authentication).

The state machine creates a user pool, a tenant specific role, and the initial user. The tenant information is stored in the tenant shard account's table as well as in the management account's table. The management account also stores a copy that is retrievable by host name. This is used by the tenant info endpoint to get the information needed for a tenant to login and to hit the appropriate GraphQL endpoint.